Text_2_SQL/requirements.txt

altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
dataclasses-json==0.6.7
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
faiss-cpu==1.11.0
filetype==1.2.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-aiplatform==1.103.0
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.25.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0

importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-vertexai==2.0.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.48
llama-index-agent-openai==0.4.12
llama-index-callbacks-arize-phoenix==0.5.1
llama-index-cli==0.4.4
llama-index-core==0.12.48
llama-index-embeddings-langchain==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.2.0
llama-index-llms-google
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
numpy==2.3.1
pandas==2.2.3
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests-toolbelt==1.0.0
s3transfer==0.13.0
setuptools==80.9.0
SQLAlchemy==2.0.41
sqlalchemy-bigquery==1.15.0
streamlit==1.46.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
